체력바 빌보드, 혈흔, 플레이어 역할 분배 및 체력바

+++ Assets/02.Scripts/Objects/Barrel.cs
@@ -6,8 +6,8 @@ using UnityEngine.Serialization;
 public class Barrel : MonoBehaviour, IDamageable
 {
     [Header("체력")]
-    [SerializeField] private int _maxHealth = 100;
-    private int _currentHealth;
+    [SerializeField] private float _maxHealth = 100;
+    private float _currentHealth;
     private bool _isDead;
     
     [Header("이펙트")]
